fix: create util folder and move error handler class
