+++ frontend/src/app/admin/admin-profile/admin-profile.ts
@@ -6,7 +6,7 @@ import { environment } from '../../../environments/environment';
 import { ResourcesService } from '../../services/resources/resources.service';
 import { Logout } from '../../auth/logout/logout';
 import { AdminEditDTO } from '../../services/admin/dto/admin-edit-dto';
-import { ErrorHandler } from '../../services/error-handler';
+import { ErrorHandler } from '../../util/error-handler';
 
 @Component({
   selector: 'app-admin-profile',
