+++ frontend/src/app/services/solidarity-fund/solidarity-fund.service.ts
@@ -4,7 +4,7 @@ import { LoginService } from '../auth/login.service';
 import { catchError, Observable, throwError } from 'rxjs';
 import { environment } from '../../../environments/environment';
 import { MessageResponse } from '../../payload/response/message-response';
-import { ErrorHandler } from '../error-handler';
+import { ErrorHandler } from '../../util/error-handler';
 
 @Injectable({
   providedIn: 'root'
